Move step ca renew to a new file.
